se añade logica de negocio y empezamos unir empleado con turno

+++ turnos/src/main/java/com/empleados/turnos/service/TurnosService.java
@@ -1,18 +1,25 @@
 package com.empleados.turnos.service;
 
+
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
+import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
+
+import com.empleados.turnos.model.Empleado;
 import com.empleados.turnos.model.Turnos;
 import com.empleados.turnos.repository.TurnoRepository;
 
 @Service
 public class TurnosService {
-    @Autowired
-    private TurnoRepository repository;
+    
+    private final TurnoRepository repository;
+
+    public TurnosService(TurnoRepository repository){
+        this.repository = repository;
+    }
 
     // Metodo Get: listar todos los turnos
     public List<Turnos> listar() {
@@ -20,39 +27,62 @@ public class TurnosService {
     }
 
     // Metodo Get: buscar por id
+    //buscamos id turno para mostrar datos del turno.
     public Turnos buscarPorId(Integer turnoId) {
-
+        //evitamos nullPointerException
+        if (turnoId == null || turnoId <= 0){
+            throw new IllegalArgumentException("El id que busca no debe ser nulo y debe ser mayor a 0");
+        }
+        //Buscamos si existe el turno por id
         Optional<Turnos> buscarTurno = repository.findById(turnoId);
-
-        if (buscarTurno == null) { // si el id no es nulo y mayor a 0 lo busca
-            throw new RuntimeException("El id no existe");
+        // si no existe lanzamos una exception
+        if (buscarTurno.isEmpty()) { // si el id no es nulo y mayor a 0 lo busca
+            throw new RuntimeException("El Turno no existe");
         }
         return buscarTurno.get();
     }
 
     // Metodo get: buscar por id del empleado
-    public Optional<Turnos> buscarIdEmpleado(Integer empleadoId) {
+    //buscamos id del empleado para mostrar los turnos de ese empleado.
+    public List<Turnos> buscarIdEmpleado(Integer empleadoId) {
 
-        if (empleadoId != null && empleadoId > 0) {
-            return repository.findByEmpleadoId(empleadoId);
+        if(empleadoId == null || empleadoId <= 0){
+            throw new IllegalArgumentException("El id del empleado no debe ser nulo y debe ser mayor a 0");
         }
-        return Optional.empty();
+        List<Turnos> buscarTurnoEmpleado = repository.findByEmpleadoId(empleadoId);
+        // isEmpty dice que no existe
+        if(buscarTurnoEmpleado.isEmpty()){
+            throw new RuntimeException("No existe un turno para ese id de empleado");
+        }
+        return buscarTurnoEmpleado;
+        
+    }
+
+    //metodo para validar datos de turnos y se va a reautilizar
+    public void validarTurno(Turnos turno){
+
+        if(turno == null){
+            throw new IllegalArgumentException("El turno no puede ser nulo");
+        }
+        if (turno.getEmpleadoId() == null || turno.getEmpleadoId() <= 0){
+            throw new IllegalArgumentException("Los datos del turno no son validos");
+        }
+        if(!"Trabajando".equals(turno.getEstado()) && !"Descansando".equals(turno.getEstado())){
+            throw new IllegalArgumentException("El estado del turno debe ser trabajando o descansando");
+        }
+
     }
 
     // Metodo post crear un turno
-    public Optional<Turnos> crearTurno(Turnos turno) {
+    public Turnos crearTurno(Turnos turno) {
         /*
          * No se valida duplicados aqui ni en la base de datos;porque el mismo empleado
          * puede tener muchos turnos de
          * trabajo. nada es unico en este microservicio excepto el id turno que lo
-         * asigna la base de datos. solo validamos que no sea nulo
-         * ni que el id del empleado dea menor a 0
+         * asigna la base de datos.
          */
-        if (turno != null && turno.getEmpleadoId() != null && turno.getEmpleadoId() > 0) {
-            repository.save(turno);
-            return Optional.of(turno);
-        }
-        return Optional.empty();
+        validarTurno(turno);
+        return repository.save(turno);
     }
 
     // Metodo Put; Actualizar un turno.
@@ -61,6 +91,8 @@ public class TurnosService {
             throw new RuntimeException("El id que desea actualizar no existe");
         }
         Optional<Turnos> existe = repository.findById(turnoId);
+
+        validarTurno(turnoNuevo); //Validamos el turno del metodo reautilizado
 
         if (existe.isPresent() && turnoNuevo != null) {
             Turnos turnos = existe.get();
@@ -76,12 +108,32 @@ public class TurnosService {
     }
 
     // Metodo delete : eliminar turno por id
-    public boolean eliminarPorId(Integer turnoId) {
+    public void  eliminarPorId(Integer turnoId) {
+
+        if(turnoId == null){
+            throw new IllegalArgumentException("El id que desea eliminar no puede ser nulo");
+        }
         if (repository.existsById(turnoId)) {
             repository.deleteById(turnoId);
-            return true;
-        }
-        throw new RuntimeException("El id que quiere eliminar no existe");
+            
+        }else{
+            throw new RuntimeException("El id que quiere eliminar no existe");
+        }  
     }
 
+    //Metodo para unir turnos con empleados solo por el id
+    //hacemos una relacion un empleado puede tener muchos turnos.
+    public Empleado obtenerTurnosConEmpleados(Integer empleadoId){
+
+       if(empleadoId == null || empleadoId <= 0){
+            throw new IllegalArgumentException("El id no debe ser nulo y debe ser mayor a 0");
+       }
+       
+       RestTemplate restTemplate = new RestTemplate();
+
+       String url = "http://localhost:8081/empleados/buscar/id/" + empleadoId;
+       Empleado empleado = restTemplate.getForObject(url, Empleado.class);
+
+       return empleado;
+    }
 }
